requirements.txt
flask~=2.3.3
git+https://gitlab.com/HumanBot000/duolingo.git@master
gevent-websocket==0.10.1